client/requirements.txt
Pillow
python-socketio

customtkinter
pyqt6
pywinstyles